code/src/requirements.txt
streamlit
faiss-cpu
pandas
numpy
sentence-transformers
google-generativeai
python-dotenv
SpeechRecognition
pyaudio